Passing match config to bots.
