+++ src/main/python/rlbot/botmanager/bot_manager_flatbuffer.py
@@ -4,12 +4,13 @@ from rlbot.messages.flat import GameTickPacket
 
 class BotManagerFlatbuffer(BotManager):
     def __init__(self, terminate_request_event, termination_complete_event, reload_request_event, bot_configuration,
-                 name, team, index, agent_class_wrapper, agent_metadata_queue, quick_chat_queue_holder):
+                 name, team, index, agent_class_wrapper, agent_metadata_queue, quick_chat_queue_holder, match_config):
         """
         See documentation on BotManager.
         """
         super().__init__(terminate_request_event, termination_complete_event, reload_request_event, bot_configuration,
-                         name, team, index, agent_class_wrapper, agent_metadata_queue, quick_chat_queue_holder)
+                         name, team, index, agent_class_wrapper, agent_metadata_queue, quick_chat_queue_holder,
+                         match_config)
         self.game_tick_flat = None
         self.game_tick_flat_binary = None
