+++ src/main/python/rlbot/botmanager/bot_manager_struct.py
@@ -9,12 +9,13 @@ from rlbot.utils.structures.rigid_body_struct import RigidBodyTick
 
 class BotManagerStruct(BotManager):
     def __init__(self, terminate_request_event, termination_complete_event, reload_request_event, bot_configuration,
-                 name, team, index, agent_class_wrapper, agent_metadata_queue, quick_chat_queue_holder):
+                 name, team, index, agent_class_wrapper, agent_metadata_queue, quick_chat_queue_holder, match_config):
         """
         See documentation on BotManager.
         """
         super().__init__(terminate_request_event, termination_complete_event, reload_request_event, bot_configuration,
-                         name, team, index, agent_class_wrapper, agent_metadata_queue, quick_chat_queue_holder)
+                         name, team, index, agent_class_wrapper, agent_metadata_queue, quick_chat_queue_holder,
+                         match_config)
         self.rigid_body_tick = None
 
     def prepare_for_run(self):
